requirements.txt
pandas = 1.3.3
pytorch = 1.2.0
sklearn